whiteboard-ai/requirements.txt
# Core ML frameworks
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
accelerate>=0.24.0

# Quantization
bitsandbytes>=0.41.0

# Computer Vision
ultralytics>=8.0.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Gradio for web interface
gradio>=4.0.0

# Utilities
numpy>=1.24.0
scipy>=1.11.0

# Training (LoRA/QLoRA)
peft>=0.7.0
trl>=0.7.0
datasets>=2.14.0

# Evaluation
editdistance>=0.6.0
